# [PERF] test_lint: skip visited docstrings