**[PERF] test_lint: compute docutils settings once**

Using `setting_override` makes docutils parse the override at every call
using its extensible argparse machinery. Changes the code to produce the
settings only once, to skip the argparse machinery.

Running the linter with crm and stock installed went fror taking 90s to
taking 60s.

---

**[PERF] test_lint: skip visited docstrings**

The methods coming with BaseModel don't have any model attached, this
made the linter re-lint every single BaseModel's method's docstring on
every model.

Running the linter with crm and stock installed went from taking 60s to
taking 3s.