@robodoo 2025-05-23
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/211466.png)](https://mergebot.odoo.com/odoo/odoo/pull/211466)